requirements.txt
bottle==0.12.25
bottle-websocket==0.2.9
cffi==1.15.1
clr-loader==0.2.6
Eel==0.16.0
future==0.18.3
gevent==23.7.0
gevent-websocket==0.10.1
greenlet==2.0.2
numpy==1.25.2
opencv-python==4.8.0.76
proxy-tools==0.1.0
pycparser==2.21
pyparsing==3.1.1
PyQRCode==1.2.1
pythonnet==3.0.2
pywebview==4.3.2
pyzbar==0.1.9
typing_extensions==4.7.1
whichcraft==0.6.1
zope.event==5.0
zope.interface==6.0